requirements.txt
Django==3.0.6
pillow==8.1.0
django-filter==2.4.0
django-materializecss-form==1.1.17
django-bootstrap-form==3.4
django-ckeditor